+++ CHANGELOG.md
@@ -6,12 +6,14 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
+- `sandstone/Panels.WizardPanel` props `nextButtonAriaLabel`, `prevButtonAriaLabel`, `noNextButton`, `noPrevButton`, and `noSteps`
 - `sandstone/Scroller` and `sandstone/VirtualList` prop `initialHiddenHeight` to provide the initially hidden height of the vertical scrollbar when `featureContent` prop in the panel set to true.
-- `sandstone/Input.InputPopup` to activate input popup without button
+- `sandstone/Input.InputPopup` component
 
 ### Fixed
 
 - `sandstone/Input.InputField` disabled colors
+- `sandstone/Panels.WizardPanel` to hide previous and next buttons appropriately
 
 ## [1.0.0-beta.1] - 2020-04-27